requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.6.0
djangorestframework==3.12.2
httpie==2.3.0
idna==2.10
oauthlib==3.1.0
py-bcrypt==0.4
pycparser==2.20
Pygments==2.7.3
PyJWT==2.0.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2020.5
razorpay==1.2.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
